Prepared socket cache
