+++ PPAndroid/app/src/main/java/edu/fcpc/polaroid/Fragment20.java
@@ -13,6 +13,8 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
 
+import edu.fcpc.polaroid.wifi.WiFiLoginHelper;
+
 /**
  * A simple {@link Fragment} subclass.
  */
