+++ PPAndroid/app/src/main/java/edu/fcpc/polaroid/wifi/WiFiHelper.java
@@ -1,22 +1,20 @@
-package edu.fcpc.polaroid;
+package edu.fcpc.polaroid.wifi;
 
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.net.wifi.WifiInfo;
 import android.net.wifi.WifiManager;
 import android.os.AsyncTask;
-import android.util.Log;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.net.InetAddress;
 import java.net.Socket;
 import java.net.UnknownHostException;
 
+import edu.fcpc.polaroid.BuildConfig;
 import edu.fcpc.polaroid.packets.PackageStatus;
 import edu.fcpc.polaroid.packets.SentPackage;
 
@@ -36,46 +34,32 @@ public abstract class WiFiHelper extends AsyncTask<String, Void, SentPackage> {
 
         try {
             WifiManager mWifiManager = (WifiManager) main.getApplicationContext().getSystemService(Context.WIFI_SERVICE);
-            //WifiInfo mWifiInfo = mWifiManager.getConnectionInfo();
-            int subnet = mWifiManager.getDhcpInfo().gateway;
-            byte[] ipBytes = new byte[]{(byte) (subnet & 0xff),
-                    (byte) (subnet >> 8 & 0xff),
-                    (byte) (subnet >> 16 & 0xff),
-                    0x00};
+            Socket socket = new Socket(SocketCache.getActiveHost(mWifiManager).getHostAddress(), 1234);
+            ObjectOutputStream objOutStream = new ObjectOutputStream(socket.getOutputStream());
 
-            for (int i = 1; i <= 254; i++) {
-                ipBytes[3] = (byte) i;
-                InetAddress address = InetAddress.getByAddress(ipBytes);
-                try {
-                    if (address.isReachable(100)) {
-                        Socket socket = new Socket(address.getHostAddress(), 1234);
-                        ObjectOutputStream objOutStream = new ObjectOutputStream(socket.getOutputStream());
+            // Send the packet to the server
+            doInBackgroundInner(objOutStream, params);
 
-                        // Send the packet to the server
-                        doInBackgroundInner(objOutStream, params);
+            // Receive message from the server
+            ObjectInputStream objInStream = new ObjectInputStream(socket.getInputStream());
+            SentPackage receivePackage = (SentPackage) objInStream.readObject();
 
-                        // Receive message from the server
-                        ObjectInputStream objInStream = new ObjectInputStream(socket.getInputStream());
-                        SentPackage receivePackage = (SentPackage) objInStream.readObject();
-
-                        return receivePackage;
-                    }
-                } catch (IOException ioe) {
-                    // TODO: The address had not opened port
-                } catch (ClassNotFoundException cnfe){
-                    // INFO: This should not enter this block
-                }
-            }
-
+            return receivePackage;
+        } catch (SocketCache.NoServerFoundException nsfe) {
             errorPackage.packageStatus = PackageStatus.NO_SERVER_FOUND;
         } catch (UnknownHostException uhe) {
             errorPackage.packageStatus = PackageStatus.HOSTNAME_NOT_FOUND;
             errorPackage.retMessage = uhe.getMessage();
+        } catch (IOException ioe) {
+            // TODO: Add exception message here
+        } catch (ClassNotFoundException cnfe) {
+            // Do nothing
         }
 
-        if(!main.getString(R.string.allownetworkbypass).toUpperCase().equals("TRUE"))
+        // Return the error message
+        if (!BuildConfig.NETWORK_BYPASS)
             return errorPackage;
-        else{
+        else {
             SentPackage networkBypass = new SentPackage();
             networkBypass.packageStatus = PackageStatus.NETWORK_BYPASS;
             return networkBypass;
@@ -87,7 +71,7 @@ public abstract class WiFiHelper extends AsyncTask<String, Void, SentPackage> {
         dialog.dismiss();
 
         // General issues
-        if (result.packageStatus == PackageStatus.HOSTNAME_NOT_FOUND){
+        if (result.packageStatus == PackageStatus.HOSTNAME_NOT_FOUND) {
             new AlertDialog.Builder(main)
                     .setTitle("Alert Box")
                     .setCancelable(false)
@@ -98,7 +82,7 @@ public abstract class WiFiHelper extends AsyncTask<String, Void, SentPackage> {
                             System.exit(0);
                         }
                     }).create().show();
-        }else if(result.packageStatus == PackageStatus.NO_SERVER_FOUND){
+        } else if (result.packageStatus == PackageStatus.NO_SERVER_FOUND) {
             new AlertDialog.Builder(main)
                     .setTitle("Alert Box")
                     .setCancelable(false)
@@ -109,8 +93,8 @@ public abstract class WiFiHelper extends AsyncTask<String, Void, SentPackage> {
                             System.exit(0);
                         }
                     }).create().show();
-        }else{
-            if(result.packageStatus == PackageStatus.NETWORK_BYPASS)
+        } else {
+            if (result.packageStatus == PackageStatus.NETWORK_BYPASS)
                 new AlertDialog.Builder(main)
                         .setTitle("Warning")
                         .setCancelable(false)
@@ -126,5 +110,6 @@ public abstract class WiFiHelper extends AsyncTask<String, Void, SentPackage> {
     }
 
     public abstract Integer doInBackgroundInner(ObjectOutputStream objOutStream, String... params) throws IOException;
+
     public abstract void onPostExecuteAfter(SentPackage sentPackage);
 }
