+++ PPAndroid/app/src/main/res/values/strings.xml
@@ -1,4 +1,3 @@
 <resources>
     <string name="app_name">Polaroid Digital Frame</string>
-    <string name="allownetworkbypass">true</string>
 </resources>
